Rename battery-notifier -> battery-monitor (mechanical, no behaviour change)
Prep for the upcoming status-file emitter (filed separately): with that
emitter the daemon stops being a pure notifier and becomes the canonical
source of battery state for any consumer on the desktop, so 'monitor'
better describes its scope. This PR is the pure rename; no behaviour
change.

Surface area:
- pkgs/battery-notifier.nix -> pkgs/battery-monitor.nix (pname,
  mainProgram, src path, comments)
- modules/services/battery-notifier.nix -> battery-monitor.nix
  (option namespace nx.services.batteryNotifier -> batteryMonitor;
  systemd.user.services.battery-notifier -> battery-monitor; unit
  Description and ExecStart; daemonConfigFile name)
- modules/services/battery-notifier/ -> battery-monitor/ (including
  nested Go-module dir battery-monitor/battery-monitor/)
- Go module path: github.com/prismatic-koi/battery-notifier ->
  battery-monitor; every import in main.go + internal/** updated;
  go mod tidy run to keep go.sum consistent
- DESIGN.md prose, AppName, log message text, comments
- flake.nix package attribute and overlay
- overlays/default.nix attribute name + comment
- modules/services/default.nix import path
- modules/services/flake-update-notifier.nix comment reference

CI (.github/workflows/pr-gate.yml) — walked carefully:
- Job names go-tests-battery-notifier -> go-tests-battery-monitor and
  nix-build-battery-notifier-checked -> nix-build-battery-monitor-checked
- pr-gate fan-in needs: list updated to the new job names
- Result-variable accessors GO_TESTS_BN_RESULT/NIX_BUILD_BN_CHECKED_RESULT
  renamed to ..._BM_... and the if-check + echos updated to match
- Path-filter globs and working-directory paths updated
- paths-filter output key renamed bn -> bm and all step 'if:' guards
  updated to match (mismatch would silently skip every step)
- Comment text refreshed

Carve-outs:
- The 'batteryNotifier.enable = false' line in the renamed module is
  the upstream openrazer NixOS module's own option name and intentionally
  stays unchanged. A NOTE comment was added so a future sed-wielding
  reader doesn't trip on it.
- Internal structured log 'event=' field values (event=startup,
  event=upower_subscribed, etc.) are observable contract and were not
  in scope per the issue body. None were touched.

Verification:
- rg battery-notifier|batteryNotifier returns only the four documented
  upstream-openrazer references
- nix build .#battery-monitor succeeds; .#battery-notifier no longer
  exists as an attribute
- nix build .#nixosConfigurations.{navi,tui}.config.system.build.toplevel
  both succeed (build log shows battery-monitor.service and
  battery-monitor-config.json being realised)
- go test ./... passes in the renamed module
- nix build with runChecks=true (homeless-shelter sandbox) passes

Refs #1974.
