+++ modules/services/battery-monitor/battery-monitor/internal/config/config_test.go
