+++ modules/services/battery-monitor/battery-monitor/internal/source/upower/upower.go
@@ -17,7 +17,7 @@ import (
 
 	"github.com/godbus/dbus/v5"
 
-	"github.com/prismatic-koi/battery-notifier/internal/state"
+	"github.com/prismatic-koi/battery-monitor/internal/state"
 )
 
 const (
