+++ modules/services/battery-monitor/battery-monitor/internal/state/state_test.go
