+++ overlays/default.nix
@@ -100,11 +100,11 @@ rec {
 
       prism = final.callPackage ../pkgs/prism.nix { };
 
-      # battery-notifier: Linux-only Go daemon (UPower + sysfs +
-      # session-bus notifications). See pkgs/battery-notifier.nix and
-      # modules/services/battery-notifier/DESIGN.md.
-      battery-notifier =
-        if final.stdenv.isLinux then final.callPackage ../pkgs/battery-notifier.nix { } else null;
+      # battery-monitor: Linux-only Go daemon (UPower + sysfs +
+      # session-bus notifications). See pkgs/battery-monitor.nix and
+      # modules/services/battery-monitor/DESIGN.md.
+      battery-monitor =
+        if final.stdenv.isLinux then final.callPackage ../pkgs/battery-monitor.nix { } else null;
 
       _macronTypePkg =
         if final.stdenv.isDarwin then final.callPackage ../pkgs/macron-type.nix { } else null;
